Extract @Nested CustomMapping test suite from Mapping into a standalone file
